+++ rfft.js
@@ -1,15 +1,3 @@
-/**
- * RFFT is a class for calculating the Discrete Fourier Transform of a signal
- * with the Fast Fourier Transform algorithm.
- *
- * This method currently only contains a forward transform but is highly optimized.
- *
- * @param {Number} bufferSize The size of the sample buffer to be computed. Must be power of 2
- * @param {Number} sampleRate The sampleRate of the buffer (eg. 44100)
- *
- * @constructor
- */
-
 import FourierTransform from "./fourier-transform.js"
 
 // lookup tables don't really gain us any speed, but they do increase
@@ -23,15 +11,30 @@ import FourierTransform from "./fourier-transform.js"
 // the rest of this was translated from C, see http://www.jjj.de/fxt/
 // this is the real split radix FFT
 
+/**
+ * RFFT is a class for calculating the Discrete Fourier Transform of a signal
+ * with the Fast Fourier Transform algorithm.
+ *
+ * This method currently only contains a forward transform but is highly optimized.
+
+ */
 export default class RFFT extends FourierTransform {
+	/**
+	 * @param {Number} bufferSize The size of the sample buffer to be computed. Must be power of 2
+	 * @param {Number} sampleRate The sampleRate of the buffer (eg. 44100)
+	 */
 	constructor(bufferSize, sampleRate) {
 		super(bufferSize, sampleRate)
-		this.trans = new Float64Array(bufferSize)
+		this.trans = new Float32Array(bufferSize)
 		this.reverseTable = new Uint32Array(bufferSize)
 		this.generateReverseTable()
 	}
 
 	// don't use a lookup table to do the permute, use this instead
+	/**
+	 *	@param {Float32Array} dest
+	 *	@param {Float32Array} source
+	 */
 	reverseBinPermute(dest, source) {
 		var bufferSize = this.bufferSize,
 			halfSize = bufferSize >>> 1,
@@ -111,6 +114,7 @@ export default class RFFT extends FourierTransform {
 	//             ...
 	// trans[n-1]   = im[1]
 
+	/** @param {Float32Array} buffer */
 	forward(buffer) {
 		var n = this.bufferSize,
 			spectrum = this.spectrum,
